+++ tests/unit/test_client_comprehensive.py
@@ -99,294 +99,6 @@ class TestNeonClientInitialization:
         assert client._cache is None
 
 
-class TestNeonClientHTTPMethods:
-    """Test NeonClient HTTP methods."""
-
-    def setup_method(self):
-        """Set up test fixtures."""
-        self.client = NeonClient(org_id="test", api_key="test")
-
-    @patch("httpx.Client.request")
-    def test_get_request_success(self, mock_request):
-        """Test successful GET request."""
-        mock_response = Mock()
-        mock_response.status_code = 200
-        mock_response.json.return_value = {"data": "test"}
-        mock_request.return_value = mock_response
-
-        result = self.client.get("/test")
-
-        assert result == {"data": "test"}
-        mock_request.assert_called_once_with(
-            "GET", "https://api.neoncrm.com/v2/test", params=None, json=None, timeout=30
-        )
-
-    @patch("httpx.Client.request")
-    def test_post_request_success(self, mock_request):
-        """Test successful POST request."""
-        mock_response = Mock()
-        mock_response.status_code = 201
-        mock_response.json.return_value = {"id": 123}
-        mock_request.return_value = mock_response
-
-        data = {"name": "test"}
-        result = self.client.post("/test", json_data=data)
-
-        assert result == {"id": 123}
-        mock_request.assert_called_once_with(
-            "POST",
-            "https://api.neoncrm.com/v2/test",
-            params=None,
-            json=data,
-            timeout=30,
-        )
-
-    @patch("httpx.Client.request")
-    def test_put_request_success(self, mock_request):
-        """Test successful PUT request."""
-        mock_response = Mock()
-        mock_response.status_code = 200
-        mock_response.json.return_value = {"updated": True}
-        mock_request.return_value = mock_response
-
-        data = {"name": "updated"}
-        result = self.client.put("/test/123", json_data=data)
-
-        assert result == {"updated": True}
-
-    @patch("httpx.Client.request")
-    def test_patch_request_success(self, mock_request):
-        """Test successful PATCH request."""
-        mock_response = Mock()
-        mock_response.status_code = 200
-        mock_response.json.return_value = {"patched": True}
-        mock_request.return_value = mock_response
-
-        data = {"field": "value"}
-        result = self.client.patch("/test/123", json_data=data)
-
-        assert result == {"patched": True}
-
-    @patch("httpx.Client.request")
-    def test_delete_request_success(self, mock_request):
-        """Test successful DELETE request."""
-        mock_response = Mock()
-        mock_response.status_code = 204
-        mock_response.json.return_value = {}
-        mock_request.return_value = mock_response
-
-        result = self.client.delete("/test/123")
-
-        assert result == {}
-
-    @patch("httpx.Client.request")
-    def test_request_with_params(self, mock_request):
-        """Test request with query parameters."""
-        mock_response = Mock()
-        mock_response.status_code = 200
-        mock_response.json.return_value = {"data": "test"}
-        mock_request.return_value = mock_response
-
-        params = {"page": 1, "size": 10}
-        self.client.get("/test", params=params)
-
-        mock_request.assert_called_once_with(
-            "GET",
-            "https://api.neoncrm.com/v2/test",
-            params=params,
-            json=None,
-            timeout=30,
-        )
-
-
-class TestNeonClientErrorHandling:
-    """Test NeonClient error handling."""
-
-    def setup_method(self):
-        """Set up test fixtures."""
-        self.client = NeonClient(org_id="test", api_key="test")
-
-    @patch("httpx.Client.request")
-    def test_400_bad_request_error(self, mock_request):
-        """Test 400 Bad Request error handling."""
-        mock_response = Mock()
-        mock_response.status_code = 400
-        mock_response.json.return_value = {"error": "Bad request"}
-        mock_request.return_value = mock_response
-
-        with pytest.raises(NeonBadRequestError):
-            self.client.get("/test")
-
-    @patch("httpx.Client.request")
-    def test_401_authentication_error(self, mock_request):
-        """Test 401 Authentication error handling."""
-        mock_response = Mock()
-        mock_response.status_code = 401
-        mock_response.json.return_value = {"error": "Unauthorized"}
-        mock_request.return_value = mock_response
-
-        with pytest.raises(NeonAuthenticationError):
-            self.client.get("/test")
-
-    @patch("httpx.Client.request")
-    def test_403_forbidden_error(self, mock_request):
-        """Test 403 Forbidden error handling."""
-        mock_response = Mock()
-        mock_response.status_code = 403
-        mock_response.json.return_value = {"error": "Forbidden"}
-        mock_request.return_value = mock_response
-
-        with pytest.raises(NeonForbiddenError):
-            self.client.get("/test")
-
-    @patch("httpx.Client.request")
-    def test_404_not_found_error(self, mock_request):
-        """Test 404 Not Found error handling."""
-        mock_response = Mock()
-        mock_response.status_code = 404
-        mock_response.json.return_value = {"error": "Not found"}
-        mock_request.return_value = mock_response
-
-        with pytest.raises(NeonNotFoundError):
-            self.client.get("/test")
-
-    @patch("httpx.Client.request")
-    def test_422_unprocessable_entity_error(self, mock_request):
-        """Test 422 Unprocessable Entity error handling."""
-        mock_response = Mock()
-        mock_response.status_code = 422
-        mock_response.json.return_value = {"error": "Validation failed"}
-        mock_request.return_value = mock_response
-
-        with pytest.raises(NeonUnprocessableEntityError):
-            self.client.get("/test")
-
-    @patch("httpx.Client.request")
-    def test_429_rate_limit_error(self, mock_request):
-        """Test 429 Rate Limit error handling."""
-        mock_response = Mock()
-        mock_response.status_code = 429
-        mock_response.headers = {"Retry-After": "60"}
-        mock_response.json.return_value = {"error": "Rate limit exceeded"}
-        mock_request.return_value = mock_response
-
-        with pytest.raises(NeonRateLimitError) as exc_info:
-            self.client.get("/test")
-
-        assert exc_info.value.retry_after == 60
-
-    @patch("httpx.Client.request")
-    def test_500_server_error(self, mock_request):
-        """Test 500 Server Error handling."""
-        mock_response = Mock()
-        mock_response.status_code = 500
-        mock_response.json.return_value = {"error": "Internal server error"}
-        mock_request.return_value = mock_response
-
-        with pytest.raises(NeonServerError):
-            self.client.get("/test")
-
-    @patch("httpx.Client.request")
-    def test_connection_error(self, mock_request):
-        """Test connection error handling."""
-        mock_request.side_effect = httpx.ConnectError("Connection failed")
-
-        with pytest.raises(NeonConnectionError):
-            self.client.get("/test")
-
-    @patch("httpx.Client.request")
-    def test_timeout_error(self, mock_request):
-        """Test timeout error handling."""
-        mock_request.side_effect = httpx.TimeoutException("Request timed out")
-
-        with pytest.raises(NeonTimeoutError):
-            self.client.get("/test")
-
-
-class TestNeonClientRetryLogic:
-    """Test NeonClient retry logic for server errors."""
-
-    def setup_method(self):
-        """Set up test fixtures."""
-        self.client = NeonClient(org_id="test", api_key="test")
-
-    @patch("httpx.Client.request")
-    @patch("time.sleep")
-    def test_retry_on_502_bad_gateway(self, mock_sleep, mock_request):
-        """Test retry logic for 502 Bad Gateway."""
-        # First call fails with 502, second succeeds
-        mock_responses = [
-            Mock(status_code=502, json=lambda: {"error": "Bad Gateway"}),
-            Mock(status_code=200, json=lambda: {"data": "success"}),
-        ]
-        mock_request.side_effect = mock_responses
-
-        result = self.client.get("/test")
-
-        assert result == {"data": "success"}
-        assert mock_request.call_count == 2
-        mock_sleep.assert_called_once()
-
-    @patch("httpx.Client.request")
-    @patch("time.sleep")
-    def test_retry_on_503_service_unavailable(self, mock_sleep, mock_request):
-        """Test retry logic for 503 Service Unavailable."""
-        mock_responses = [
-            Mock(status_code=503, json=lambda: {"error": "Service Unavailable"}),
-            Mock(status_code=200, json=lambda: {"data": "success"}),
-        ]
-        mock_request.side_effect = mock_responses
-
-        result = self.client.get("/test")
-
-        assert result == {"data": "success"}
-        assert mock_request.call_count == 2
-
-    @patch("httpx.Client.request")
-    @patch("time.sleep")
-    def test_retry_on_504_gateway_timeout(self, mock_sleep, mock_request):
-        """Test retry logic for 504 Gateway Timeout."""
-        mock_responses = [
-            Mock(status_code=504, json=lambda: {"error": "Gateway Timeout"}),
-            Mock(status_code=200, json=lambda: {"data": "success"}),
-        ]
-        mock_request.side_effect = mock_responses
-
-        result = self.client.get("/test")
-
-        assert result == {"data": "success"}
-        assert mock_request.call_count == 2
-
-    @patch("httpx.Client.request")
-    @patch("time.sleep")
-    def test_retry_exhausted(self, mock_sleep, mock_request):
-        """Test when all retries are exhausted."""
-        # All calls fail with 502
-        mock_request.return_value = Mock(
-            status_code=502, json=lambda: {"error": "Bad Gateway"}
-        )
-
-        with pytest.raises(NeonServerError):
-            self.client.get("/test")
-
-        # Should try 3 times (initial + 2 retries)
-        assert mock_request.call_count == 3
-        assert mock_sleep.call_count == 2
-
-    @patch("httpx.Client.request")
-    def test_no_retry_on_client_errors(self, mock_request):
-        """Test that client errors (4xx) are not retried."""
-        mock_request.return_value = Mock(
-            status_code=400, json=lambda: {"error": "Bad Request"}
-        )
-
-        with pytest.raises(NeonBadRequestError):
-            self.client.get("/test")
-
-        # Should only try once
-        assert mock_request.call_count == 1
-
-
 class TestNeonClientCaching:
     """Test NeonClient caching functionality."""
 
@@ -417,59 +129,6 @@ class TestNeonClientCaching:
 
         # Should not raise error
         client.clear_cache()
-
-
-class TestNeonClientHeaders:
-    """Test NeonClient request headers."""
-
-    def setup_method(self):
-        """Set up test fixtures."""
-        self.client = NeonClient(org_id="test_org", api_key="test_key")
-
-    def test_default_headers(self):
-        """Test that default headers are set correctly."""
-        headers = self.client._get_default_headers()
-
-        assert "Authorization" in headers
-        assert "Content-Type" in headers
-        assert "User-Agent" in headers
-        assert "NEON-ORG-ID" in headers
-
-        assert headers["Content-Type"] == "application/json"
-        assert headers["NEON-ORG-ID"] == "test_org"
-        assert "neon-crm-sdk" in headers["User-Agent"]
-
-    def test_authorization_header(self):
-        """Test that authorization header is properly encoded."""
-        headers = self.client._get_default_headers()
-        auth_header = headers["Authorization"]
-
-        assert auth_header.startswith("Basic ")
-
-        # Decode and verify
-        import base64
-
-        encoded_creds = auth_header.split(" ")[1]
-        decoded_creds = base64.b64decode(encoded_creds).decode()
-        assert decoded_creds == "test_org:test_key"
-
-    @patch("httpx.Client.request")
-    def test_headers_sent_in_request(self, mock_request):
-        """Test that headers are included in requests."""
-        mock_response = Mock()
-        mock_response.status_code = 200
-        mock_response.json.return_value = {"data": "test"}
-        mock_request.return_value = mock_response
-
-        self.client.get("/test")
-
-        # Verify the request was made with proper headers
-        call_args = mock_request.call_args
-        assert "headers" in call_args.kwargs
-
-        headers = call_args.kwargs["headers"]
-        assert "Authorization" in headers
-        assert "NEON-ORG-ID" in headers
 
 
 class TestNeonClientContextManager:
